Need to fixed negative values and carry

+++ Vm/zjump.c
@@ -6,7 +6,7 @@
 /*   By: ahamouda <ahamouda@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/06/13 11:18:50 by ahamouda          #+#    #+#             */
-/*   Updated: 2016/08/30 16:50:57 by ahamouda         ###   ########.fr       */
+/*   Updated: 2016/09/01 23:16:48 by ahamouda         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,7 +33,7 @@ void		zjump(t_vm_data *arena, t_proc *process)
 		display_zjump_instruction(process, arena);
 	if (!process->carry)
 		return ;
-	tmp = get_param_value(arena, process->pc + 1, get_n_param_size(arena, 1));
+	tmp = (short)get_param_value(arena, process->pc + 1, get_n_param_size(arena, 1));
 	if ((arena->verbosity & 8) == 8)
 		ft_printf_fd(arena->fd, " It jumped by %d bytes !\n", tmp);
 	process->pc = (process->pc + (tmp % IDX_MOD)) % arena->mem_size;
